renames Bar to ProgressBar
