+++ index.js
@@ -11,7 +11,7 @@ const Newline = require('./lib/components/newline');
 const Indent = require('./lib/components/indent');
 const Group = require('./lib/components/group');
 const Text = require('./lib/components/text');
-const Bar = require('./lib/components/bar');
+const ProgressBar = require('./lib/components/progress-bar');
 
 exports.StringComponent = StringComponent;
 exports.Component = Component;
@@ -20,7 +20,7 @@ exports.Newline = Newline;
 exports.Indent = Indent;
 exports.Group = Group;
 exports.Text = Text;
-exports.Bar = Bar;
+exports.ProgressBar = ProgressBar;
 
 const noop = () => {};
