+++ lib/components/progress-bar.js
